+++ packages/config/src/Currency.ts
@@ -1,15 +1,25 @@
-import type { CurrencyProps } from '.';
+import type { Type } from '.';
 
-export const Currency = (config: CurrencyProps): CurrencyProps => {
-	return {
-		code: config?.code || 'USD',
-		singularLabel: config?.singularLabel || 'Dollar',
-		pluralLabel: config?.pluralLabel || 'Dollars',
-		sign: config?.sign || '$',
-		signB4: config?.signB4,
-		decimalPlaces: config?.decimalPlaces >= 0 ? config.decimalPlaces : 2,
-		decimalMark: config?.decimalMark || '.',
-		thousandsSeparator: config?.thousandsSeparator || ',',
-		subunits: config?.subunits >= 0 ? config.subunits : Math.pow(10, config?.decimalPlaces || 2),
-	};
+export const Currency = (config?: Partial<Type.Currency>): Type.Currency => {
+	const { subunits: exclude, ...defaults } = defaultCurrency;
+	const subunits = Subunits(config);
+	return { ...defaults, ...config, subunits };
+};
+
+function Subunits(config?: Partial<Type.Currency>): Type.Currency['subunits'] {
+	if (config?.subunits) return config.subunits;
+	const dp = config?.decimalPlaces ?? defaultCurrency.decimalPlaces;
+	return Math.pow(10, dp);
+}
+
+const defaultCurrency: Type.Currency = {
+	code: 'USD',
+	singularLabel: 'Dollar',
+	pluralLabel: 'Dollars',
+	sign: '$',
+	signB4: false,
+	decimalPlaces: 2,
+	decimalMark: '.',
+	thousandsSeparator: ',',
+	subunits: 100, // Math.pow(10, 2);
 };
