+++ packages/config/src/index.ts
@@ -4,7 +4,4 @@ export * from './Locale';
 export * from './SiteUrl';
 export * from './Timezone';
 
-export * from './useConfig';
-export * from './useConfigData';
-
 export * from './types';
