--- packages/config/src/useConfigData.ts
@@ -1,52 +1,0 @@
-import { useMemoStringify } from '@eventespresso/hooks';
-import { useMemo } from 'react';
-import { Currency, DateTimeFormats, Locale, LocaleProps, SiteUrl, Timezone, ConfigDataProps } from './';
-
-export const useConfigData = (): ConfigDataProps => {
-	const api = useMemoStringify(window?.eventEspressoData?.api);
-	const config = useMemoStringify(window?.eventEspressoData?.config);
-
-	return useMemo(
-		() => ({
-			brandName: config?.coreDomain?.brandName || 'Event Espresso',
-			currency: Currency(config?.siteCurrency),
-			currentUser: config?.currentUser,
-			generalSettings: config?.generalSettings,
-			dateTimeFormats: DateTimeFormats({
-				dateFormat: config?.generalSettings?.dateFormat,
-				timeFormat: config?.generalSettings?.timeFormat,
-			}),
-			locale: Locale({
-				site: config?.locale?.site || '',
-				siteTimezone: config?.locale?.siteTimezone || {}, // TODO:
-				user: config?.locale?.user || '',
-			}),
-			nonce: api?.restApiNonce || '',
-			sitePermissions: config?.sitePermissions || [],
-			siteUrl: SiteUrl({
-				admin: config?.siteUrls?.admin || '',
-				home: config?.siteUrls?.home || '',
-			}),
-			timezone: Timezone({
-				city: config?.locale?.siteTimezone?.city || '',
-				name: config?.locale?.siteTimezone?.name || '',
-				offset: config?.locale?.siteTimezone?.offset || 0,
-			}),
-			wp_debug: config?.wp_debug || false,
-		}),
-		[
-			api?.restApiNonce,
-			config?.coreDomain?.brandName,
-			config?.currentUser,
-			config?.generalSettings,
-			config?.locale?.site,
-			config?.locale?.siteTimezone,
-			config?.locale?.user,
-			config?.siteCurrency,
-			config?.sitePermissions,
-			config?.siteUrls?.admin,
-			config?.siteUrls?.home,
-			config?.wp_debug,
-		]
-	);
-};
