+++ packages/types/src/index.ts
@@ -1,3 +1,2 @@
-export * from './window-types';
 export * from './generic-types';
 export * from './data';
